Workaround: minimize.py hangs while processing kernel/locking/qspinlock.c after v4.4
In kernel/locking/qspinlock.c, the same C file qspinlock.c is internally included from the original C file,
which causes misbehavior of minimize.py when processing "#include" lines in C files.

This workaround is not to minimize qspinlock.c but exceptionally just copy the file as it is,
because in such case qspinlock.c cannot be uniquely minimized due to internal "#include" in the same file.

As far as I examined, it was only qspinlock.c that internally includes himself.

Signed-off-by: Kotaro Hashimoto <kotaro.hashimoto.jv@hitachi.com>